as.data.frame method for VariableCatalog

This method gives you a view of a VariableCatalog as a data.frame in order to facilitate further exploration.

# S3 method for VariableCatalog
as.data.frame(x, row.names = NULL,
  optional = FALSE, keys = c("alias", "name", "type"), ...)

Arguments

x
A VariableCatalog, as produced by variables(ds).

row.names
part of as.data.frame signature. Ignored.

optional
part of as.data.frame signature. Ignored.

keys
A character vector of the variable catalog attributes which you would like included in the data.frame. To include all attributes or see which ones are available, set keys to "all". By default, the function will return three fields: c("alias", "name", "type").

...
Additional arguments passed to data.frame

Value

A data.frame including metadata about each variable stored in the variable catalog. The fields in the data.frame match the keys argument provided to the function, and each row represents a variable.

Details

Modifying the data.frame produced by this function will not update the dataset on the Crunch server. Other methods exist for updating the metadata in the variable catalog. See vingette("variables", package = "crunch").

Examples

# NOT RUN {
ds <- loadDataset("iris")
vars <- variables(ds)
var_df <- as.data.frame(vars, keys = "all")
# }