Organize Variables within a Dataset

Variables in the Crunch web application can be viewed in an ordered, hierarchical list. These objects and methods allow you to modify that order from R.

Details

A VariableOrder object is a subclass of list that contains VariableGroups. VariableGroup objects contain a group name and an set of "entities", which can be variable references or other nested VariableGroups.

Slots

group

character, the name of the VariableGroup. In the constructor and more generally, this field can be referenced as "name" as well.

entities

a character vector of variable URLs, or a list containing a combination of variable URLs and VariableGroup objects.

duplicates

logical: should duplicate variable references be allowed in this object? Default is FALSE.

vars

either NULL or a VariableCatalog. If not NULL, it will be used to look up variable names from the URLs.