Extract and modify Crunch objects

Extract and modify Crunch objects

This function is no-op because the items are already updated on the server with other methods called prior to it.

# S4 method for class 'AbstractCategories,ANY,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'AbstractCategories,character,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'AbstractCategories,numeric,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'AbstractCategories,character,ANY,ANY'
x[i, j, ...] <- value

# S4 method for class 'AbstractCategories,character'
x[[i, j, ...]]

# S4 method for class 'AbstractCategories,character,ANY,ANY'
x[[i, j, ...]] <- value

# S4 method for class 'AbstractCategory'
x$name

# S4 method for class 'AbstractCategory'
x$name <- value

# S4 method for class 'ScriptCatalog,numeric'
x[[i, j, ...]]

# S4 method for class 'Categories,ANY,ANY,ANY'
x[i, j, ...] <- value

# S3 method for class 'CrunchDataFrame'
x[i, j, drop = TRUE]

# S3 method for class 'CrunchDataFrame'
x[i, j] <- value

# S3 method for class 'CrunchDataFrame'
x[[i]]

# S3 method for class 'CrunchDataFrame'
x[[i]] <- value

# S3 method for class 'CrunchDataFrame'
x$i

# S3 method for class 'CrunchDataFrame'
x$i <- value

# S4 method for class 'CubeDims,ANY,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'CrunchCube,ANY,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'TransformsList,ANY,missing,NULL'
x[[i, j]] <- value

# S4 method for class 'DatasetCatalog,numeric'
x[[i, j, ...]]

# S4 method for class 'DatasetCatalog,character,missing,DatasetTuple'
x[[i, j]] <- value

# S4 method for class 'CrunchDataset,ANY,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'CrunchDataset,logical,missing'
x[i, j, ..., drop = FALSE]

# S4 method for class 'CrunchDataset,character,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'CrunchDataset,VariableGroup,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'CrunchDataset,VariableOrder,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'CrunchDataset,missing,ANY'
x[i, j, ..., drop = FALSE]

# S4 method for class 'CrunchDataset,CrunchLogicalExpr,missing'
x[i, j, ..., drop = FALSE]

# S4 method for class 'CrunchDataset,CrunchLogicalExpr,ANY'
x[i, j, ..., drop = FALSE]

# S4 method for class 'CrunchDataset'
subset(x, ...)

# S4 method for class 'CrunchDataset,ANY'
x[[i, ..., drop = FALSE]]

# S4 method for class 'CrunchDataset,character'
x[[i, ..., drop = FALSE]]

# S4 method for class 'CrunchDataset'
x$name

# S4 method for class 'CrunchDataset,character,missing,CrunchVariable'
x[[i]] <- value

# S4 method for class 'CrunchDataset,ANY,missing,CrunchVariable'
x[[i]] <- value

# S4 method for class 'CrunchDataset,character,missing,ANY'
x[[i]] <- value

# S4 method for class 'CrunchDataset,ANY,ANY,ANY'
x[[i]] <- value

# S4 method for class 'CrunchDataset,character,missing,NULL'
x[[i]] <- value

# S4 method for class 'CrunchDataset,ANY,missing,NULL'
x[[i]] <- value

# S4 method for class 'CrunchDataset'
x$name <- value

# S4 method for class 'CrunchDataset,character,missing,CrunchGeography'
x[[i]] <- value

# S4 method for class 'CrunchDataset,ANY,missing,CrunchGeography'
x[[i]] <- value

# S4 method for class 'CrunchDataset,ANY,missing,list'
x[i, j] <- value

# S4 method for class 'CrunchDataset,ANY,missing,CrunchDataset'
x[i, j] <- value

# S4 method for class 'CrunchDataset,CrunchExpr,ANY,ANY'
x[i, j] <- value

# S4 method for class 'ShojiCatalog,character,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'ShojiCatalog,numeric,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'ShojiCatalog,logical,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'ShojiCatalog,ANY,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'ShojiCatalog,ANY'
x[[i, j, ...]]

# S4 method for class 'ShojiCatalog,character'
x[[i, j, ...]]

# S4 method for class 'ShojiCatalog'
x$name

# S4 method for class 'ShojiCatalog'
x$name <- value

# S4 method for class 'ShojiCatalog,ANY,missing,ShojiCatalog'
x[i, j] <- value

# S4 method for class 'DeckCatalog,ANY'
x[[i, j, ...]]

# S4 method for class 'DeckCatalog,character'
x[[i, j, ...]]

# S4 method for class 'DeckCatalog,ANY,missing,CrunchDeck'
x[[i, j]] <- value

# S4 method for class 'CrunchDeck,ANY,missing,AnalysisCatalog'
x[[i, j]] <- value

# S4 method for class 'SlideCatalog,ANY,missing,CrunchSlide'
x[[i, j]] <- value

# S4 method for class 'AnalysisCatalog,ANY,missing,Analysis'
x[[i, j]] <- value

# S4 method for class 'CrunchDeck,ANY'
x[[i, j, ...]]

# S4 method for class 'CrunchDeck,ANY,ANY,ANY'
x[[i, j]] <- value

# S4 method for class 'CrunchExpr,CrunchLogicalExpr,ANY'
x[i, j, ..., drop = FALSE]

# S4 method for class 'CrunchExpr,logical,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'CrunchExpr,numeric,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'FilterCatalog,numeric'
x[[i, j, ...]]

# S4 method for class 'FilterCatalog,character,missing,CrunchLogicalExpr'
x[[i, j]] <- value

# S4 method for class 'FilterCatalog,numeric,missing,CrunchLogicalExpr'
x[[i, j]] <- value

# S4 method for class 'FilterCatalog,character,missing,CrunchFilter'
x[[i, j]] <- value

# S4 method for class 'FilterCatalog,numeric,missing,CrunchFilter'
x[[i, j]] <- value

# S4 method for class 'MemberCatalog,character'
x[[i, j, ...]]

# S4 method for class 'MemberCatalog,ANY,missing,ANY'
x[[i, j]] <- value

# S4 method for class 'MemberCatalog,character,missing,NULL'
x[[i, j]] <- value

# S4 method for class 'MultitableCatalog,numeric'
x[[i, j, ...]]

# S4 method for class 'MultitableCatalog,character,missing,formula'
x[[i, j]] <- value

# S4 method for class 'MultitableCatalog,numeric,missing,formula'
x[[i, j]] <- value

# S4 method for class 'MultitableCatalog,ANY,missing,Multitable'
x[[i, j]] <- value

# S4 method for class 'MultitableCatalog,ANY,missing,NULL'
x[[i, j]] <- value

# S4 method for class 'PermissionCatalog,character'
x[[i, j, ...]]

# S4 method for class 'ProjectFolder,character,missing,list'
x[[i, j]] <- value

# S4 method for class 'ProjectFolder,character,missing,ProjectFolder'
x[[i, j]] <- value

# S4 method for class 'ShojiFolder,numeric'
x[[i, ..., drop = FALSE]]

# S4 method for class 'ShojiFolder,character'
x[[i, ..., drop = FALSE]]

# S4 method for class 'ShojiOrder,ANY,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'ShojiOrder,character,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'ShojiOrder,ANY'
x[[i, j, ...]]

# S4 method for class 'ShojiOrder,character'
x[[i, j, ...]]

# S4 method for class 'ShojiOrder'
x$name

# S4 method for class 'ShojiOrder,character,missing,ShojiOrder'
x[i, j] <- value

# S4 method for class 'ShojiOrder,ANY,missing,ShojiOrder'
x[i, j] <- value

# S4 method for class 'ShojiOrder,character,missing,list'
x[[i, j]] <- value

# S4 method for class 'ShojiOrder,character,missing,character'
x[[i, j]] <- value

# S4 method for class 'ShojiOrder,character,missing,OrderGroup'
x[[i, j]] <- value

# S4 method for class 'ShojiOrder,ANY,missing,OrderGroup'
x[[i, j]] <- value

# S4 method for class 'ShojiOrder,ANY,missing,ANY'
x[[i, j]] <- value

# S4 method for class 'ShojiOrder,ANY,missing,NULL'
x[[i, j]] <- value

# S4 method for class 'ShojiOrder,character,missing,NULL'
x[[i, j]] <- value

# S4 method for class 'ShojiOrder,character,missing,ShojiOrder'
x[[i, j]] <- value

# S4 method for class 'ShojiOrder'
x$name <- value

# S4 method for class 'OrderGroup,ANY,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'OrderGroup,character,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'OrderGroup,character'
x[[i, j, ...]]

# S4 method for class 'OrderGroup,ANY'
x[[i, j, ...]]

# S4 method for class 'OrderGroup'
x$name

# S4 method for class 'OrderGroup,character,missing,list'
x[[i, j]] <- value

# S4 method for class 'OrderGroup,character,missing,character'
x[[i, j]] <- value

# S4 method for class 'OrderGroup,character,missing,ShojiOrder'
x[[i, j]] <- value

# S4 method for class 'OrderGroup,character,missing,OrderGroup'
x[[i, j]] <- value

# S4 method for class 'OrderGroup,ANY,missing,OrderGroup'
x[[i, j]] <- value

# S4 method for class 'OrderGroup,numeric,missing,NULL'
x[[i, j]] <- value

# S4 method for class 'OrderGroup,character,missing,NULL'
x[[i, j]] <- value

# S4 method for class 'OrderGroup'
x$name <- value

# S4 method for class 'SlideCatalog,ANY'
x[[i, j, ...]]

# S4 method for class 'SlideCatalog,numeric,missing,CrunchSlide'
x[[i, j]] <- value

# S4 method for class 'CrunchAnalysisSlide,ANY'
x[[i, j, ...]]

# S4 method for class 'CrunchAnalysisSlide,ANY,ANY,ANY'
x[[i, j]] <- value

# S4 method for class 'AnalysisCatalog,ANY'
x[[i, j, ...]]

# S4 method for class 'AnalysisCatalog,numeric,missing,formula'
x[[i, j]] <- value

# S4 method for class 'AnalysisCatalog,numeric,missing,Analysis'
x[[i, j]] <- value

# S4 method for class 'AnalysisCatalog,numeric,missing,list'
x[[i, j]] <- value

# S4 method for class 'Subvariables,character'
x[[i, j, ...]]

# S4 method for class 'Subvariables,numeric'
x[[i, j, ...]]

# S4 method for class 'Subvariables,character,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'Subvariables,character,missing,CrunchVariable'
x[[i]] <- value

# S4 method for class 'Subvariables,ANY,missing,CrunchVariable'
x[[i]] <- value

# S4 method for class 'Subvariables,ANY,missing,NULL'
x[[i]] <- value

# S4 method for class 'Subvariables,ANY,missing,ANY'
x[[i]] <- value

# S4 method for class 'Subvariables,character,missing,Subvariables'
x[i] <- value

# S4 method for class 'Subvariables,ANY,missing,Subvariables'
x[i] <- value

# S4 method for class 'Subvariables,ANY,missing,ANY'
x[i] <- value

# S4 method for class 'ArrayVariable,character,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'ArrayVariable,missing,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'ArrayVariable,missing,character'
x[i, j, ..., drop = TRUE]

# S4 method for class 'ArrayVariable,ANY'
x[[i, j, ...]]

# S4 method for class 'ArrayVariable,character'
x[[i, j, ...]]

# S4 method for class 'ArrayVariable'
x$name

# S4 method for class 'ArrayVariable,ANY,missing,ANY'
x[[i]] <- value

# S4 method for class 'ArrayVariable,character,missing,ANY'
x[[i]] <- value

# S4 method for class 'ArrayVariable'
x$name <- value

# S4 method for class 'TabBookResult,numeric'
x[[i, j, ...]]

# S4 method for class 'TabBookResult,character'
x[[i, j, ...]]

# S4 method for class 'MultitableResult,ANY'
x[[i, j, ...]]

# S4 method for class 'TeamCatalog,numeric'
x[[i, j, ...]]

# S4 method for class 'TeamCatalog,character,missing,list'
x[[i, j]] <- value

# S4 method for class 'TeamCatalog,character,missing,CrunchTeam'
x[[i, j]] <- value

# S4 method for class 'UserCatalog,character,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'UserCatalog,character'
x[[i, j, ...]]

# S4 method for class 'VariableCatalog,numeric'
x[[i, j, ...]]

# S4 method for class 'VariableCatalog,character,missing,VariableTuple'
x[[i, j]] <- value

# S4 method for class 'VariableCatalog,character,missing,CrunchVariable'
x[[i, j]] <- value

# S4 method for class 'VariableCatalog,VariableOrder,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'VariableCatalog,VariableGroup,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'VariableCatalog,VariableOrder,missing,VariableCatalog'
x[i, j] <- value

# S4 method for class 'VariableCatalog,VariableGroup,missing,VariableCatalog'
x[i, j] <- value

# S4 method for class 'VariableOrder,character,missing,CrunchDataset'
x[[i, j]] <- value

# S4 method for class 'VariableGroup,character,missing,CrunchDataset'
x[[i, j]] <- value

# S4 method for class 'CrunchVariable,ANY,missing,ANY'
x[i, j] <- value

# S4 method for class 'CrunchVariable,ANY,missing,NULL'
x[i, j] <- value

# S4 method for class 'TextVariable,ANY,missing,character'
x[i, j] <- value

# S4 method for class 'NumericVariable,ANY,missing,numeric'
x[i, j] <- value

# S4 method for class 'DatetimeVariable,ANY,missing,Date'
x[i, j] <- value

# S4 method for class 'DatetimeVariable,ANY,missing,POSIXt'
x[i, j] <- value

# S4 method for class 'CrunchVariable,ANY,missing,CrunchExpr'
x[i, j] <- value

# S4 method for class 'CrunchVariable,ANY,missing,CrunchVariable'
x[i, j] <- value

# S4 method for class 'CrunchVariable,ANY,missing,CrunchLogicalExpr'
x[i, j] <- value

# S4 method for class 'CategoricalVariable,ANY,missing,numeric'
x[i, j] <- value

# S4 method for class 'CategoricalVariable,ANY,missing,character'
x[i, j] <- value

# S4 method for class 'CategoricalVariable,ANY,missing,factor'
x[i, j] <- value

# S4 method for class 'CategoricalArrayVariable,ANY,missing,numeric'
x[i, j] <- value

# S4 method for class 'CategoricalArrayVariable,ANY,missing,character'
x[i, j] <- value

# S4 method for class 'CategoricalArrayVariable,ANY,missing,factor'
x[i, j] <- value

# S4 method for class 'CrunchVariable,ANY,missing,logical'
x[i, j] <- value

# S4 method for class 'CrunchVariable,ANY'
is.na(x) <- value

# S4 method for class 'CrunchVariable,CrunchExpr,ANY'
x[i, j, ..., drop = FALSE]

# S4 method for class 'CrunchVariable,numeric,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'CrunchVariable,logical,ANY'
x[i, j, ..., drop = TRUE]

Arguments

x

a Crunch object (Dataset, Variable, CrunchExpr, Catalog, CrunchCube, etc.)

i

The elements to extract; as with R, this can generally be (1) a logical vector of length matching x; (2) a character vector of appropriate length, which can generally be URLs or also names, aliases, or other identifier; (3) integer indices, potentially negative, to select. Datasets and Variables can also be subset by CrunchExpr. Note that when subsetting Datasets and variables by R vectors, the input i is turned into a CrunchExpr so that it can be used in API queries.

j

For two-dimensional objects, elements to take from the columnar dimension. Note that Datasets work like data.frames in that you can extract columns either by ds[, j] or by the list-like ds[i] (with no comma).

...

additional arguments supported by some methods. For example, some catalogs support a secondary vector of elements to match against like emails(x) or owners(x); by default this is names(x). This lets (for example) you to look up variables by URL but fall back to name.

drop

Ignored and always assumed FALSE

value

For updating, an object of the appropriate class and size to insert. This is generally the same class of object you would get from the extract method; i.e. for x[[i]] <- value, value must be the same class as x[[i]]. Exceptions include ds$var <- VariableDefinition(...) to create a new variable, and ds$var[i] <- value, which can take R vectors and CrunchExpr.

name

for $, the same as i for [[

Value

[ generally returns a subset of x, same class and "self" URL; when subsetting with a CrunchExpr, the expression is recorded in an attribute of the object. [[ and $ return appropriate extractions from x, wherever possible without making an additional API request. The assignment methods return x appropriately updated. Most assignment methods do not make an API request except where they clearly are used to create a new entity (as in assigning a variable definition to a Dataset); for editing metadata attributes, these methods generally assume that the API request to set the changes on the server happens in a different method (e.g. in names(variables(ds)[1:5]) <- value, assignment happens in the names<- method).