Methods on Cube objects

These methods provide an array-like interface to the CrunchCube object.

# S4 method for CubeDims
dimnames(x)

# S4 method for CubeDims
dim(x)

# S4 method for CubeDims
is.na(x)

# S4 method for CrunchCube
dimensions(x)

# S4 method for CrunchCube,CubeDims
dimensions(x) <- value

# S4 method for CubeDims,ANY,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for CrunchCube
dim(x)

# S4 method for CrunchCube
dimnames(x)

# S4 method for CrunchCube,ANY,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for CrunchCube
measures(x)

# S4 method for CrunchCube
names(x)

# S4 method for CrunchCube
aliases(x)

# S4 method for CrunchCube
descriptions(x)

# S4 method for CrunchCube
types(x)

# S4 method for CrunchCube
notes(x)

Arguments

x
a CrunchCube or its CubeDims component.

value
for dimensions<- a CubeDims object to overwrite a CrunchCube dimensions

i
used with [ to extract a dimension

j
not used

...
not used

drop
not used

Value

Generally, the same shape of result that each of these functions return when applied to an array object.

See also