Delete a Crunch object from the server

These methods delete entities, notably Datasets and Variables within them, from the server. This action is permanent and cannot be undone, so it should not be done lightly. Consider instead using archive for datasets and hide for variables.

# S4 method for CrunchDataset
delete(x, ...)

# S4 method for Multitable
delete(x, ...)

# S4 method for CrunchProject
delete(x, ...)

# S4 method for ShojiObject
delete(x, ...)

# S4 method for ANY
delete(x, ...)

# S4 method for CrunchTeam
delete(x, ...)

# S4 method for CrunchVariable
delete(x, ...)

# S4 method for VariableTuple
delete(x, ...)

Arguments

x
a Crunch object

...
additional arguments, in the generic

Details

Deleting requires confirmation. In an interactive session, you will be asked to confirm. To avoid that prompt, or to delete objects from a non-interactive session, wrap the call in with_consent to give your permission to delete.

See also

hide deleteDataset