Hide/Unhide or Privatize/Deprivatize Variables

Both hidden and private are hidden from most views in crunch by default. Hidden variables can be accessed by an user, while private variables (and all variables derived from them) are only accessible by users granted "editor" access to the dataset and so can be used to secure personally identifiable information from non-editors of a dataset.

hiddenFolder(x)

privateFolder(x)

hide(x)

unhide(x)

privatize(x)

deprivatize(x)

# S4 method for CrunchDataset
hiddenFolder(x)

# S4 method for VariableCatalog
hiddenFolder(x)

# S4 method for VariableFolder
hiddenFolder(x)

# S4 method for CrunchVariable
hide(x)

# S4 method for VariableCatalog
hide(x)

# S4 method for CrunchVariable
unhide(x)

# S4 method for VariableCatalog
unhide(x)

hideVariables(dataset, variables)

hiddenVariables(x) <- value

unhideVariables(dataset, variables)

hiddenVariables(dataset, key = namekey(dataset))

# S4 method for CrunchDataset
privateFolder(x)

# S4 method for VariableCatalog
privateFolder(x)

# S4 method for VariableFolder
privateFolder(x)

# S4 method for CrunchVariable
privatize(x)

# S4 method for VariableCatalog
privatize(x)

# S4 method for CrunchVariable
deprivatize(x)

# S4 method for VariableCatalog
deprivatize(x)

privatise(x)

deprivatise(x)

privatizeVariables(dataset, variables)

privatiseVariables(dataset, variables)

privateVariables(x) <- value

deprivatizeVariables(dataset, variables)

deprivatiseVariables(dataset, variables)

privateVariables(dataset, key = namekey(dataset))

Arguments

x
a Variable, VariableCatalog, or dataset to hide/unhide/privatize/deprivatize

dataset
A dataset

variables
Variables to change status of

value
Replacement values for assignment methods.

key
(for hiddenVariables() / privateVariables() the Variable attribute to return. Default is "alias", following getOption("crunch.namekey.dataset").

Details

There are several ways to assign variables into these categories and access them:

hideVariables() / privatizeVariables() - take a character vector of variable aliases and makes them hidden/private. (unhideVariables() / deprivatizeVariables() put them back in the main variable catalog).

hide() / privatize() - take a CrunchVariable or VariableCatalog and make them hidden/private. (unhide() / deprivatize() put them back in the main variable catalog).

hiddenFolder() / privateFolder() - take a dataset and return a folder that contains the hidden/private variables. This folder is like other CrunchFolders and so you can use mkdir() to create subfolders and mv() to move them in/out.

hiddenVariables() / privateVariabiles() - return a character vector of variables that are hidden/private. You can assign into the catalog to add variables or assign to NULL to remove all of them.