is.na for Categories

Crunch categorical variables allow you to set multiple categories as missing. For instance, you might have "not answered" and "doesn't know" both coded as missing. This function returns a logical vector of all dataset entries that fall into any of the missing categories. It also allows you to append additional categories to the list of missing categories using the setter.

# S4 method for Categories
is.na(x)

# S4 method for Categories,character
is.na(x) <- value

# S4 method for Categories,logical
is.na(x) <- value

# S4 method for Category
is.na(x)

# S4 method for Category,logical
is.na(x) <- value

Arguments

x
Categories or a single Category

value
To change the missingness of categories, supply either:

a logical vector of equal length of the categories (or length 1 for the Category method); or

the names of the categories to mark as missing. If supplying the latter, any categories already indicated as missing will remain missing.

Value

Getters return logical, a named vector in the case of the Categories method; setters return x duly modified.