Find an entity in an order object

Find an entity in an order object

locateEntity(x, ord)

Arguments

x
Variable or Dataset, depending on the type of order, or URL for it

ord
ShojiOrder (VariableOrder or DatasetOrder)

Value

If x is found in ord, a character vector of group names that provide the "path" to the entity. The length of the vector corresponds to the depth of nesting. If not found, NA is returned