Convert a child class of Insertion into a proper Insertion

The Crunch API expects that Subtotals, Headings, and other insertions all have the same shape. Before sending insertions to the server, we need to call makeInsertion on any insertions to make sure they are proper Insertions. This process sometimes requires a variable's categories object (e.g. in order to convert from category names to category ids)

# S4 method for Subtotal
makeInsertion(x, var_categories)

# S4 method for Heading
makeInsertion(x, var_categories)

# S4 method for Insertion
makeInsertion(x, var_categories)

# S4 method for SummaryStat
makeInsertion(x, var_categories)

Arguments

x
an object the is a child of Insertion (e.g. Subtotal, Heading)

var_categories
categories (from categories(variable)) to used by makeInsertions to make Insertions from a child object (e.g. Subtotal or Heading)

Value

an Insertion object