Retry

Retry an expression. This is useful for situations where a web resource is not yet available. You can set options("crunch_retry_wait" = X) some number larger than the default 0.1 in your script if you are working with large exports.

retry(
  expr,
  wait = getOption("crunch_retry_wait", default = 0.1),
  max.tries = 10
)

Arguments

expr
An expression

wait
The time in seconds to wait before retrying the expression. Defaults to 0.1.

max.tries
The number of times to retry the expression