Extract and modify subsets of Catalog-type objects
Source: R/dataset-catalog.R, R/filters.R, R/members.R, and 9 more

Extract and modify subsets of Catalog-type objects

# S4 method for DatasetCatalog,numeric
[[(x, i, j, ...)

# S4 method for DatasetCatalog,character,missing,DatasetTuple
[[(x, i, j) <- value

# S4 method for FilterCatalog,numeric
[[(x, i, j, ...)

# S4 method for FilterCatalog,character,missing,CrunchLogicalExpr
[[(x, i,
  j) <- value

# S4 method for FilterCatalog,numeric,missing,CrunchLogicalExpr
[[(x, i,
  j) <- value

# S4 method for FilterCatalog,character,missing,CrunchFilter
[[(x, i, j) <- value

# S4 method for FilterCatalog,numeric,missing,CrunchFilter
[[(x, i, j) <- value

# S4 method for MemberCatalog,character
[[(x, i, j, ...)

# S4 method for MemberCatalog,ANY,missing,ANY
[[(x, i, j) <- value

# S4 method for MemberCatalog,character,missing,`NULL`
[[(x, i, j) <- value

# S4 method for MultitableCatalog,numeric
[[(x, i, j, ...)

# S4 method for MultitableCatalog,character,missing,formula
[[(x, i, j) <- value

# S4 method for MultitableCatalog,numeric,missing,formula
[[(x, i, j) <- value

# S4 method for MultitableCatalog,ANY,missing,`NULL`
[[(x, i, j) <- value

# S4 method for PermissionCatalog,character
[[(x, i, j, ...)

# S4 method for ProjectCatalog,numeric
[[(x, i, j, ...)

# S4 method for ProjectCatalog,character,missing,list
[[(x, i, j) <- value

# S4 method for ProjectCatalog,character,missing,CrunchProject
[[(x, i,
  j) <- value

# S4 method for ShojiCatalog,character,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for ShojiCatalog,numeric,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for ShojiCatalog,logical,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for ShojiCatalog,ANY,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for ShojiCatalog,ANY
[[(x, i, j, ...)

# S4 method for ShojiCatalog,character
[[(x, i, j, ...)

# S4 method for ShojiCatalog
$(x, name)

# S4 method for ShojiCatalog
$(x, name) <- value

# S4 method for ShojiCatalog,ANY,missing,ShojiCatalog
[(x, i, j) <- value

# S4 method for ShojiFolder,numeric
[[(x, i, ..., drop = FALSE)

# S4 method for ShojiFolder,character
[[(x, i, ..., drop = FALSE)

# S4 method for TeamCatalog,numeric
[[(x, i, j, ...)

# S4 method for TeamCatalog,character,missing,list
[[(x, i, j) <- value

# S4 method for TeamCatalog,character,missing,CrunchTeam
[[(x, i, j) <- value

# S4 method for UserCatalog,character,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for UserCatalog,character
[[(x, i, j, ...)

# S4 method for VariableCatalog,numeric
[[(x, i, j, ...)

# S4 method for VariableCatalog,character,missing,VariableTuple
[[(x, i,
  j) <- value

# S4 method for VariableCatalog,character,missing,CrunchVariable
[[(x, i,
  j) <- value

# S4 method for VariableCatalog,VariableOrder,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for VariableCatalog,VariableGroup,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for VariableCatalog,VariableOrder,missing,VariableCatalog
[(x, i,
  j) <- value

# S4 method for VariableCatalog,VariableGroup,missing,VariableCatalog
[(x, i,
  j) <- value

Arguments

x
a Catalog object

i
which catalog elements to extract

j
Invalid

...
additional arguments (for example secondary a vector of elements to match against like emails(x) or owners(x) by default this is names(x))

value
For updating, an object of the appropriate class and size to insert

drop
Invalid

name
for $, the same as i for [[

Value

A subset of x if extracting, otherwise x duly modified